fixing import/no-cycle issue and export issue
