+++ src/data/DDItem.tsx
@@ -9,25 +9,8 @@ import {
 import DDTable from './DDTable';
 import DDChart from './DDChart';
 import { ErrorText } from '../components/ErrorText';
-
-type DDPayload = { [key: string]: any };
-type DDPassProps = { [key: string]: any };
-
-export interface DDComponentProps {
-  payload: DDPayload;
-  title: string;
-  props?: DDPassProps;
-}
-
-export interface DDItemProps {
-  url: string;
-  fullSize?: boolean;
-  title: string;
-  dataComponentProps?: DDPassProps;
-  Component: (props: DDComponentProps) => JSX.Element;
-  SkeletonComponent: () => JSX.Element;
-  HeaderSkeletonComponent?: () => JSX.Element;
-}
+import type { DDPayload } from './types';
+import { DDItemProps } from './types';
 
 interface DDErrorProps {
   error: Error;
@@ -65,11 +48,7 @@ export default function DDItem({
           </GridItemContent>
         </>
       ) : (
-        <Component
-          payload={payload as DDPayload}
-          title={title}
-          props={dataComponentProps}
-        />
+        <Component payload={payload} title={title} props={dataComponentProps} />
       )}
     </GridItem>
   );
